Change search and search highlight color settings
The color settings for the search highlight (both incremental and the
highlight of all the matches after pressing enter) were unreadable with
the color scheme that I otherwise like (slate). So, I changed it to
something more readable and turned off hlsearch for now, because I find
it distracting to have all the search results permanently highlighted
until doing :noh or another search.