Add bootstrap script and update mac setup scripts
bootstrap.sh is the starting point which can be downloaded and executed
directly from github (since the repo is public). That will create an ssh
key, prompt me to add it to github, install the xcode cli tooling, and
then finally clone this repo via ssh.

From there I would change to the ~/projects/dotfiles/setup/mac directory
and run `./setup.sh` to install all the things.

At the point of this commit I'm still in a hybrid setup where `rake
install` is doing a lot, after which you would run this. Eventually this
will be the place for all the setup and we'll get rid of the ruby/rake
dependency and just use shell scripts for all the things.