Install and activate versions of each programming language
Not having node, python, or ruby installed in particular cause problems
with other things like vim plugins that rely on them, so it's nice to
have some version installed and configured out of the box. I decided to
hard-code the versions and just note in the README that they could be
updated prior to running setup, or additional versions could be
installed after the fact.