Remove eslint ls, comment out prettier, disable tsserver formatting
I feel like I'd like to have eslint running, but as it's currently
configured it gives me duplicate error messages between that and
tsserver. I'm not sure what the solution is there, as we often have
eslint rules for a project that we want to know about, which may or may
not be different than what tsserver is telling us.

Also, prettier with null-ls doesn't seem to be respecting the local_only
flag. It was working but again today it stopped and auto-formatted some
javascript in a repo that didn't have prettier installed. I think I'm
going to switch to formatting.nvim (or try it out at least) and
hopefully get a bit more easy control over that stuff. We'll see.

And yeah, by default tsserver does auto-formatting, so if you don't
explicitly turn those capabilities off, you'll get that formatting too
I think.